Move symtab stuff out of service.c and into its own file.
