+++ x86_64/service.c
@@ -3,7 +3,6 @@
 #include <pci.h>
 #include <virtio.h>
 #include <tfs.h>
-#include <elf64.h>
 
 extern void startup();
 extern void start_interrupts();
@@ -114,75 +113,12 @@ void attach_storage(heap h, heap virtual, tuple root, block_read r, block_write 
     runloop();
 }
 
-static rtrie elf_symtable;
-
-CLOSURE_0_4(elf_symtable_add, void, char *, u64, u64, u8);
-void elf_symtable_add(char * name, u64 a, u64 len, u8 info)
-{
-    int type = ELF64_ST_TYPE(info);
-
-    /* store bind info? */
-    if (a == 0 || len == 0 || !name || name[0] == '\0' ||
-	(type != STT_FUNC && type != STT_OBJECT))
-	return;
-
-    assert(elf_symtable);
-
-    char * m;
-    if ((m = rtrie_lookup(elf_symtable, a, 0)) ||
-	(m = rtrie_lookup(elf_symtable, a + len - 1, 0))) {
-#ifdef ELF_SYMTAB_DEBUG
-	rprintf("!!! %s (%P) exists in rtrie as \"%s\"; skipping\n",
-		name, a, m);
-#endif
-	return;
-    }
-
-    rtrie_insert(elf_symtable, a, len, name);
-}
-
-char * find_elf_sym(u64 a, u64 *offset)
-{
-    if (!elf_symtable)
-	return 0;
-
-    range r;
-    char * m = rtrie_lookup(elf_symtable, a, &r);
-    if (!m)
-	return 0;
-    
-    if (offset)
-	*offset = a - r.start;
-
-    return m;
-}
-
-void add_elf_syms(heap h, buffer b)
-{
-    if (elf_symtable)
-	elf_symbols(b, closure(h, elf_symtable_add));
-    else
-	console("can't add ELF symbols; symtab not initialized\n");
-}
-
-void init_service_new_stack(heap pages, heap physical, heap backed, heap virtual)
+static void read_kernel_syms(heap h, heap virtual, heap pages)
 {
     u64 kern_base = INVALID_PHYSICAL;
     u64 kern_length;
 
-    // just to find maintain the convention of faulting on zero references
-    map(0, INVALID_PHYSICAL, PAGESIZE, pages);
-
-    heap misc = allocate_rolling_heap(backed, 8);
-    //    misc = debug_heap(misc, misc); 
-    runqueue = allocate_queue(misc, 64);
-    start_interrupts(pages, misc, physical);
-    init_extra_prints();
-    init_runtime(misc);
-
-    /* set up elf symbol table for better backtrace info */
-    elf_symtable = rtrie_create(misc);
-
+    /* add kernel symbols */
     for_regions(e)
 	if (region_type(e) == REGION_KERNIMAGE) {
 	    kern_base = region_base(e);
@@ -194,13 +130,28 @@ void init_service_new_stack(heap pages, heap physical, heap backed, heap virtual
 	    rprintf("kernel ELF image at %P, length %P, mapped at %P\n",
 		    kern_base, kern_length, v);
 #endif
-	    add_elf_syms(misc, alloca_wrap_buffer(v, kern_length));
+	    add_elf_syms(h, alloca_wrap_buffer(v, kern_length));
 	}
     
     if (kern_base == INVALID_PHYSICAL) {
 	console("kernel elf image region not found; no debugging symbols\n");
     }
+}
 
+void init_service_new_stack(heap pages, heap physical, heap backed, heap virtual)
+{
+    // just to find maintain the convention of faulting on zero references
+    map(0, INVALID_PHYSICAL, PAGESIZE, pages);
+
+    heap misc = allocate_rolling_heap(backed, 8);
+    //    misc = debug_heap(misc, misc); 
+    runqueue = allocate_queue(misc, 64);
+    start_interrupts(pages, misc, physical);
+    init_extra_prints();
+    init_runtime(misc);
+    init_symtab(misc);
+    read_kernel_syms(misc, virtual, pages);
+    
     tuple root = allocate_tuple();
     initialize_timers(misc);
     init_pci(misc);
