+++ x86_64/symtab.h
@@ -1,0 +1,4 @@
+void init_symtab(heap h);
+void add_elf_syms(heap h, buffer b);
+char * find_elf_sym(u64 a, u64 *offset);
+
