+++ x86_64/x86_64.h
@@ -158,6 +158,8 @@ void print_frame(context f);
 void *load_elf(buffer elf, u64 offset, heap pages, heap bss);
 void elf_symbols(buffer elf, closure_type(each, void, char *, u64, u64, u8));
 
+#include <symtab.h>
+
 #define mov_to_cr(__x, __y) __asm__("mov %0,%%"__x: :"a"(__y):);
 #define mov_from_cr(__x, __y) __asm__("mov %%"__x", %0":"=a"(__y):);
